pyproject.toml
[build-system]
requires = ["setuptools==75.8.0", "wheel==0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "forge"
version = "0.1.0"
description = "Forge phase-one MVP for ML training data pipelines"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pytest==8.3.5",
  "pyyaml==6.0.2",
]

[project.optional-dependencies]
s3 = ["boto3==1.37.0"]
logging = ["structlog==25.1.0"]
lance = ["lance==1.2.1", "pyarrow==19.0.0"]
serve = ["torch==2.6.0", "matplotlib==3.10.0"]

[project.scripts]
forge = "cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"
exclude = ["^tests/"]
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["boto3", "boto3.*", "structlog", "lance", "pyarrow", "torch", "matplotlib", "matplotlib.*"]
ignore_missing_imports = true
